+++ inst/include/TreeTools/ClusterTable.h
@@ -23,13 +23,20 @@
   (b) = S[--Spos];                                               \
   (a) = S[--Spos]
 
+// Required by TreeDist 2.9.2
+// TODO Remove in later version, to prefer ct_stack_size
 #define CT_STACK_SIZE 4
 
 #define CT_IS_LEAF(a) (a) <= n_tip
 
+// Required by TreeDist 2.9.2
+// TODO Remove in later version, to prefer ct_max_leaves
 const int_fast32_t CT_MAX_LEAVES = 16383;
 
 namespace TreeTools {
+
+  constexpr int_fast32_t ct_max_leaves = 16383;
+  constexpr int_fast32_t ct_stack_size = 4;
 
   class ClusterTable {